Refactor/extract more logic from index to helper
