+++ .gitignore
@@ -1,1 +1,3 @@
 /node_modules
+users_qa.json
+user_staging.json
